fix the hiding of eaf browser buffer's view on Windows
Eaf browser buffer's view hides sometimes when the eaf on Windows is called by
emacs on WSL. Eaf browser runs well but on the background and we can't see it.
Apparently, the reparent doesn't work correctly for some reason. The workaround
for now is to call eaf-monitor-configuration-change to make the eaf browser
buffer's view appear.