Make install-eaf.sh run in eaf directory
The install-eaf.sh script should run in the eaf directory. This commit fixes that for linux and osx. As I can not test on Windows, I have not touched that command.

btw. the npm install errors with the following message (although the browser still works):
```
 ERROR  Failed to compile with 2 errors                                                                                                                                                                                            10:19:22 AM

This dependency was not found:

* @fortawesome/fontawesome-svg-core in ./node_modules/@fortawesome/vue-fontawesome/index.es.js, ./src/main.js
```
Let me know if I should open an issue for this, and if you need more information about it. I am on Fedora 34.